Renamed net.yakclient to dev.extframework across all files
This major commit changes the name of the entire hierarchy from net.yakclient to dev.extframework. This includes renaming packages in many kotlin files as well as the changes in URLs used in the project. Also the task names, classes, and other references related to "yakclient" were updated to "extframework". Some redundant code was removed. This completes the rebranding of yakclient to extframework in the codebase.